fix: requested changes

+++ crates/common/beacon_api_types/src/sync.rs
@@ -5,7 +5,8 @@ use ssz_derive::{Decode, Encode};
 pub struct SyncStatus {
     #[serde(with = "serde_utils::quoted_u64")]
     pub head_slot: u64,
-    pub sync_distance: usize,
+    #[serde(with = "serde_utils::quoted_u64")]
+    pub sync_distance: u64,
     pub is_syncing: bool,
     pub is_optimistic: bool,
     pub el_offline: bool,
